Move subclass into component_tree dir
